bound rtmNative to eventing, implemented rtmNative callback promise resolution, and keyed commitId by deviceId

+++ FitpaySDK/PaymentDevice/Interfaces/MockPaymentDeviceInterface.swift
@@ -42,8 +42,8 @@ public class MockPaymentDeviceInterface : NSObject, PaymentDeviceBaseInterface {
     }
     
     public func sendAPDUData(data: NSData, sequenceNumber: UInt16) {
-        var score: UInt = 0x90000000000000
-        let data = NSData(bytes: &score, length: sizeof(UInt))
+        var score: Double = 0x90000000000000
+        let data = NSData(bytes: &score, length: sizeof(Double))
         let packet = ApduResultMessage(msg: data)
         
         if let apduResponseHandler = self.paymentDevice.apduResponseHandler {
